Remove over-eager ftplugin indent settings
It's not really appropriate for filetype plugins to set indent-related options.

In this case, the appropriate options (autoindent, expandtab, and sw=2)
are already being set in indent/slim.vim.

smartindent is meant for C-like languages where the normal C indentation
doesn't work. There's no reason to use it here (autoindent is what you
want, but that's already set in indent/slim.vim).

There's no reason to set all of tabstop, softtabstop, and shiftwidth;
and when using smarttab, only shiftwidth is required.

However, smarttab is a global option with no buffer-local value, so
setting it (even with setlocal) changes the option for all buffers. Thus
it should be left up to the user.